Update include assertion docs for object properties
Originally added in #230, asserting that an object contains a subset of
properties and values was never documented. Questions like #193 have
popped up asking for this feature even though it's been present since
1.9.0.

The discussion in #193 focused mostly around getting something like
this to work via the `property` assertion, but that discussion should
be considered moot by the existing functionality already present in
`include` that was simply overlooked.

Resolves: #193